Moved routes/reviews.py routes to api/salons/reviews.py and deleted old routes/reviews.py
